Minor changes to fix issues with replaying
When replaying an animation the completed state is never reached as the
property is not reset after `play()` is called subsequent times. I've
changed the `play()` call to reset this value to `false`.

I also looked into the issues reported for Promises and realized we
could do a better job when resetting it. This is where I identified the
previous problem.

I do not believe any API changes will break and I've reviewed the
examples and everything looks fine.

Closes the following issues:

- https://github.com/juliangarnier/anime/issues/373
- https://github.com/juliangarnier/anime/issues/378
- https://github.com/juliangarnier/anime/issues/392

And the following redundant PR:

- https://github.com/juliangarnier/anime/pull/408